requirements.txt
pafy==0.5.4
ffmpeg-python==0.2.0
discord.py==1.2.3
discord==1.0.1
python-dotenv==0.10.3
PyNaCl==1.3.0
youtube-dl
